requirements.txt
python-box==5.1.1
tika==1.19
requests_cache==0.5.2
numpy==1.17.3
requests==2.22.0
pygbif==0.4.0
neo4j==1.7.6
ete3==3.1.1
apache_airflow==1.10.10
pandas==0.25.3
pytest==3.10.1
rdflib==4.2.2
SPARQLWrapper==1.8.5
docker_py==1.10.6
python_magic==0.4.18
pynomer==0.1.1
PyYAML==5.3.1
